+++ km3pipe/utils/h5info.py
@@ -30,13 +30,29 @@ def h5info(h5name, raw=False):
     with tables.open_file(h5name) as h5:
         att_list = h5.root._v_attrs._f_list()
         if raw:
+            print("HDF5 Meta Data")
+            print("--------------")
             print(h5.root._v_attrs.__repr__())
+            if '/header' in h5:
+                print("\nHeader ('/header')")
+                print("----------------")
+                print(h5.get_node('/header')._v_attrs.__repr__())
             return
         if not att_list and not raw:
             print("Sorry, no metadata.")
+        else:
+            print("HDF5 Meta Data")
+            print("--------------")
         for att in h5.root._v_attrs._f_list():
             ver = h5.root._v_attrs[att]
             print("{}: {}".format(att, ver))
+        if '/header' in h5:
+            print("\nHeader (/header)")
+            print("----------------")
+            node = h5.get_node('/header')
+            for att in node._v_attrs._f_list():
+                ver = node._v_attrs[att]
+                print("{}: {}".format(att, ver))
 
 
 def main():