+++ km3pipe/tests/test_tools.py
@@ -8,7 +8,7 @@ import numpy as np
 from km3pipe.testing import TestCase, StringIO
 from km3pipe.tools import (
     unpack_nfirst, split, namedtuple_with_defaults, remain_file_pointer,
-    decamelise, camelise, bincenters, issorted, lstrip, chunks, is_coherent)
+    decamelise, camelise, issorted, lstrip, chunks, is_coherent)
 
 __author__ = "Tamas Gal"
 __copyright__ = "Copyright 2016, Tamas Gal and the KM3NeT collaboration."
@@ -158,10 +158,6 @@ class TestCamelCaseConverter(TestCase):
 
 
 class TestMisc(TestCase):
-    def test_binlims(self):
-        bins = np.linspace(0, 20, 21)
-        assert bincenters(bins).shape[0] == bins.shape[0] - 1
-
     def test_issorted(self):
         assert issorted([1, 2, 3])
         assert not issorted([2, 3, 1])