+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Unreleased changes
 * ``km3pipe.plot``: Common plotting helpers
 * A handful utility functions for ``km3pipe.shell.Script``, like ``cp``,
   ``iget``...
+* ``kp.tools.bincenters`` now lives in ``kp.plot``. 
 
 Version 7
 ---------