+++ CHANGELOG.rst
@@ -4,6 +4,21 @@ Unreleased changes
 Version 7
 ---------
 
+7.16.0 / 2018-01-28
+~~~~~~~~~~~~~~~~~~~
+
+* ``km3pipe.plot``: Common plotting helpers
+* A handful utility functions for ``km3pipe.shell.Script``, like ``cp``,
+  ``iget``...
+* ``kp.tools.bincenters`` now lives in ``kp.plot``. 
+* ``kp.db.DBManager.trigger_setup`` can now retrieve trigger setups for a given
+  OID
+* Add ``n_digits`` option in ``kp.io.evt.EvtPump`` for file number index
+  when iterating over multiple files.
+* ``kp.math`` has some helpers for bootstrapping confidence intervals
+  when fitting probability distributions via max LLH (in scipy.stats)
+* Docs: move statistics examples to own section, show some distribution fits
+
 7.15.0 / 2018-01-19
 ~~~~~~~~~~~~~~~~~~~
 * ``TimeslicePump`` now supports the readout of any stream ("L0", "L1", "SN"...)